pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mokuro-bunko"
version = "0.1.0"
description = "A WebDAV server for Mokuro reader with shared manga library, per-user reading progress, and optional OCR processing"
readme = "README.md"
license = "MPL-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Gnathonic", email = "gnathonic@gmail.com" }
]
keywords = ["mokuro", "webdav", "manga", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = [
    "wsgidav>=4.0",
    "cheroot>=10.0",
    "pyyaml>=6.0",
    "bcrypt>=4.0",
    "watchdog>=3.0",
    "Pillow>=10.0",
    "click>=8.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "httpx>=0.25",
    "playwright>=1.40",
    "ruff>=0.1",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
    "shiv>=1.0",
]

[project.scripts]
mokuro-bunko = "mokuro_bunko.__main__:main"

[project.urls]
Homepage = "https://github.com/Gnathonic/mokuro-bunko"
Repository = "https://github.com/Gnathonic/mokuro-bunko"
Issues = "https://github.com/Gnathonic/mokuro-bunko/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mokuro_bunko"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/config.example.yaml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["mokuro_bunko"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
mypy_path = "src"
packages = ["mokuro_bunko"]

[[tool.mypy.overrides]]
module = [
    "wsgidav.*",
    "cheroot.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/mokuro_bunko"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
